feat(models): rename multi_categoricals module to multi_discretes and move multi_embeddings into it.
