+++ tests/sample/models/components/test_multi_discretes.py
@@ -1,7 +1,63 @@
 import pytest
 import torch
+from torch.distributions import Categorical
+from torch.distributions.distribution import Distribution
 
-from sample.models.components.multi_embeddings import MultiEmbeddings
+from sample.models.components.multi_discretes import (
+    FCMultiCategoricalHead,
+    MultiCategoricals,
+    MultiEmbeddings,
+)
+
+
+class TestMultiCategoricals:
+    @pytest.fixture
+    def distributions(self) -> list[Categorical]:
+        choices_per_dist = [3, 2, 5]
+        batch_size = 8
+        return [
+            Categorical(logits=torch.zeros(batch_size, c)) for c in choices_per_dist
+        ]
+
+    @pytest.fixture
+    def multi_categoricals(self, distributions) -> MultiCategoricals:
+        return MultiCategoricals(distributions)
+
+    def test_init(self, multi_categoricals: MultiCategoricals):
+        assert multi_categoricals.batch_shape == (8, 3)
+
+    def test_sample(self, multi_categoricals: MultiCategoricals):
+        assert multi_categoricals.sample().shape == (8, 3)
+        assert multi_categoricals.sample((1, 2)).shape == (1, 2, 8, 3)
+
+    def test_log_prob(self, multi_categoricals: MultiCategoricals):
+        sampled = multi_categoricals.sample()
+        assert multi_categoricals.log_prob(sampled).shape == sampled.shape
+
+    def test_entropy(self, multi_categoricals: MultiCategoricals):
+        assert multi_categoricals.entropy().shape == (8, 3)
+
+
+class TestFCMultiCategoricalHead:
+    @pytest.mark.parametrize(
+        """
+        batch,
+        dim_in,
+        choices_per_category,
+        """,
+        [
+            (8, 256, [3, 3, 3, 2, 2]),
+            (1, 16, [1, 2, 3, 4, 5]),
+        ],
+    )
+    def test_fc_multi_categorical_head(self, batch, dim_in, choices_per_category):
+        policy = FCMultiCategoricalHead(dim_in, choices_per_category)
+        input = torch.randn(batch, dim_in)
+        dist = policy(input)
+        assert isinstance(dist, Distribution)
+        assert dist.sample().shape == (batch, len(choices_per_category))
+        assert dist.log_prob(dist.sample()).shape == (batch, len(choices_per_category))
+        assert dist.entropy().shape == (batch, len(choices_per_category))
 
 
 class TestMultiEmbeddings:
